requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
python-dotenv>=1.0
psycopg2-binary>=2.9
openai>=1.0
